backend/requirements.txt
fastapi>=0.109,<1
uvicorn[standard]>=0.27
python-dotenv>=1.0
huggingface_hub>=0.27
pytest>=8.0
httpx>=0.27
